API for Species
