+++ django_project/frontend/urls.py
@@ -38,7 +38,6 @@ from frontend.api_views.property import (
     UpdatePropertyInformation,
     UpdatePropertyBoundaries,
     PropertyDetail,
-    Properties,
 )
 from frontend.api_views.upload import (
     BoundaryFileUpload,
@@ -46,9 +45,6 @@ from frontend.api_views.upload import (
     BoundaryFileList,
     BoundaryFileSearch,
     BoundaryFileSearchStatus
-)
-from frontend.api_views.species import (
-    TaxonViewSet
 )
 
 urlpatterns = [
@@ -99,11 +95,6 @@ urlpatterns = [
         PropertyList.as_view(),
         name='property-list'
     ),
-    path(
-        'api/property/',
-        Properties.as_view(),
-        name='properties'
-    ),
     re_path(
         r'^api/property/detail/(?P<id>\d+)/?$',
         PropertyDetail.as_view(),
@@ -153,11 +144,6 @@ urlpatterns = [
         'switch-organisation/<int:organisation_id>/',
         switch_organisation,
         name='switch-organisation'
-    ),
-    path(
-        'api/species/',
-        TaxonViewSet.as_view(),
-        name='species'
     ),
     path('map/', MapView.as_view(), name='map'),
     path('help/', HelpView.as_view(), name='help'),
