+++ django_project/species/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from .views import TaxonView
+
+urlpatterns = [
+    path('species/',TaxonView.as_view(),name='species'),
+    ]
